requirements.txt
amqp==2.5.0
beautifulsoup4==4.8.1
billiard==3.6.0.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.1.8
django-cors-middleware==1.3.1
djangorestframework==3.9.3
djangorestframework-simplejwt==4.3.0
EbookLib==0.17.1
html5lib==1.0.1
idna==2.8
kombu==4.6.3
lxml==4.3.3
mysqlclient==1.4.2.post1
numpy==1.16.4
pandas==0.24.2
PyJWT==1.7.1
python-dateutil==2.8.0
python-docx==0.8.10
python-Levenshtein==0.12.0
python-memcached==1.59
pytz==2018.9
redis==3.2.1
requests==2.22.0
rules==2.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.4
urllib3==1.25.6
vine==1.3.0
webencodings==0.5.1
